server/requirements.txt
flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
python-dotenv==1.0.0
# Uncomment and install this manually if needed for PostgreSQL
# psycopg2-binary==2.9.6
PyJWT==2.6.0
flask-bcrypt==1.0.1
Flask-JWT-Extended==4.5.2